Bloc2 - Data Architecture/Projet/Architecture/fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
neo4j==5.14.1
clickhouse-driver==0.2.6

# Kafka
confluent-kafka==2.3.0
kafka-python==2.0.2

# ML & Data Science
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
mlflow==2.9.2

# Storage
boto3==1.29.7
minio==7.2.0

# Monitoring & Logging
elasticsearch==8.11.0
python-json-logger==2.0.7

# Utilities
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
